improvements

+++ nautilus-web/src/main/java/thiagodnf/nautilus/web/controller/ExecutionController.java
@@ -1,13 +1,15 @@
 package thiagodnf.nautilus.web.controller;
 
-import java.util.Date;
 import java.util.List;
+
+import javax.validation.Valid;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -21,13 +23,15 @@ import thiagodnf.nautilus.core.encoding.NSolution;
 import thiagodnf.nautilus.core.normalize.AbstractNormalize;
 import thiagodnf.nautilus.core.objective.AbstractObjective;
 import thiagodnf.nautilus.core.util.SolutionUtils;
+import thiagodnf.nautilus.plugin.extension.ProblemExtension;
+import thiagodnf.nautilus.web.dto.ExecutionSettingsDTO;
 import thiagodnf.nautilus.web.model.Execution;
-import thiagodnf.nautilus.web.model.GenerateIndicator;
-import thiagodnf.nautilus.web.model.Parameters;
-import thiagodnf.nautilus.web.model.Settings;
+import thiagodnf.nautilus.web.model.User;
 import thiagodnf.nautilus.web.service.ExecutionService;
 import thiagodnf.nautilus.web.service.FlashMessageService;
 import thiagodnf.nautilus.web.service.PluginService;
+import thiagodnf.nautilus.web.service.SecurityService;
+import thiagodnf.nautilus.web.util.Messages;
 
 @Controller
 @RequestMapping("/execution/{executionId:.+}")
@@ -39,30 +43,29 @@ public class ExecutionController {
 	private ExecutionService executionService;
 	
 	@Autowired
+	private SecurityService securityService;
+	
+	@Autowired
 	private PluginService pluginService;
 	
 	@Autowired
 	private FlashMessageService flashMessageService;
 	
 	@GetMapping("")
-	public String view(Model model, 
-			@PathVariable("executionId") String executionId) {
+	public String view(@PathVariable String executionId, Model model) {
 		
 		LOGGER.info("Displaying '{}'", executionId);
 		
-		Execution execution = executionService.findById(executionId);
-		Parameters parameters = execution.getParameters();
-		Settings settings = execution.getSettings();
+		Execution execution = executionService.findExecutionById(executionId);
 		
-		String pluginId = parameters.getPluginId();
-		String problemId = parameters.getProblemId();
+		ProblemExtension problemExtension = pluginService.getProblemById(execution.getProblemId());
 		
-		AbstractNormalize normalizer = pluginService.getNormalizers().get(settings.getNormalizeId());
-		AbstractDuplicatesRemover duplicatesRemover = pluginService.getDuplicatesRemovers().get(settings.getDuplicatesRemoverId());
-		AbstractColorize colorizer = pluginService.getColorizers().get(settings.getColorizeId());
-		AbstractCorrelation correlation = pluginService.getCorrelationers().get(settings.getCorrelationId());
+		AbstractNormalize normalizer = pluginService.getNormalizers().get(execution.getNormalizeId());
+		AbstractDuplicatesRemover duplicatesRemover = pluginService.getDuplicatesRemovers().get(execution.getDuplicatesRemoverId());
+		AbstractColorize colorizer = pluginService.getColorizers().get(execution.getColorizeId());
+		AbstractCorrelation correlation = pluginService.getCorrelationers().get(execution.getCorrelationId());
 		
-		List<AbstractObjective> objectives = pluginService.getObjectivesByIds(pluginId, problemId, parameters.getObjectiveIds());
+		List<AbstractObjective> objectives = problemExtension.getObjectives();
 		
 		List<NSolution<?>> solutions = execution.getSolutions();
 		
@@ -71,8 +74,7 @@ public class ExecutionController {
 		List<NSolution<?>> colorfulSolutions = colorizer.execute(distinctSolutions);
 		
 		model.addAttribute("correlations", correlation.execute(objectives, normalizedSolutions));
-		model.addAttribute("plugin", pluginService.getPluginWrapper(pluginId));
-		model.addAttribute("problem", pluginService.getProblemExtension(pluginId, problemId));
+		model.addAttribute("problem", problemExtension);
 		model.addAttribute("objectives", objectives);
 		model.addAttribute("solutions", colorfulSolutions);
 		model.addAttribute("execution", execution);
@@ -81,9 +83,7 @@ public class ExecutionController {
 		model.addAttribute("colorizers", pluginService.getColorizers());
 		model.addAttribute("correlationers", pluginService.getCorrelationers());
 		model.addAttribute("reducers", pluginService.getReducers());
-		model.addAttribute("settings", execution.getSettings());
-		model.addAttribute("indicators", pluginService.getIndicatorFactory(pluginId));
-		model.addAttribute("generateIndicator", new GenerateIndicator());
+		model.addAttribute("executionSettingsDTO", executionService.convertToExecutionSettingsDTO(execution));
 		
 		return "execution";
 	}
@@ -93,49 +93,48 @@ public class ExecutionController {
 			RedirectAttributes ra,
 			@PathVariable("executionId") String executionId) {
 
-		Execution execution = executionService.findById(executionId);
+		Execution execution = executionService.findExecutionById(executionId);
 
-		executionService.deleteById(executionId);
-
-		Parameters parameters = execution.getParameters();
-
-		flashMessageService.success(ra, "msg.delete.execution.success", execution.getId());
+		User user = securityService.getLoggedUser().getUser(); 
 		
-		return "redirect:/plugin/" + parameters.getPluginId() + "/#executions";
+		if(execution.getUserId().equalsIgnoreCase(user.getId())) {
+			
+			executionService.deleteById(executionId);
+			
+			flashMessageService.success(ra, Messages.EXECUTION_DELETE_SUCCESS, execution.getId());
+		}else {
+			flashMessageService.error(ra, Messages.EXECUTION_DELETE_FAIL_NO_OWNER);
+		}
+		
+		return "redirect:/home/";
 	}
 	
 	@PostMapping("/duplicate")
-	public String duplicate(Model model,
+	public String duplicate(
+			@PathVariable("executionId") String executionId,
 			RedirectAttributes ra,
-			@PathVariable("executionId") String executionId) {
+			Model model) {
 
-		Execution execution = executionService.findById(executionId);
-
-		execution.setId(null);
-		execution.setDate(new Date());
+		executionService.duplicate(executionId);
 		
-		executionService.save(execution);
-
-		flashMessageService.success(ra, "msg.duplicate.execution.success", execution.getId());
+		flashMessageService.success(ra, Messages.EXECUTION_DUPLICATE_SUCCESS);
 		
-		return "redirect:/plugin/" + execution.getParameters().getPluginId() + "#executions";
+		return "redirect:/home/";
 	}
 	
-	@PostMapping("/save/settings")
-	public String saveSettings(Model model,
-			RedirectAttributes ra,
-			@PathVariable("executionId") String executionId, 
-			Settings settings) {
+	@PostMapping("/settings/save")
+	public String saveSettings(
+			@PathVariable("executionId") String executionId,
+			@Valid ExecutionSettingsDTO executionSettingsDTO,
+			BindingResult bindingResult,
+			Model model,
+			RedirectAttributes ra) {
 
-		Execution execution = executionService.findById(executionId);
+		executionService.updateSettings(executionId, executionSettingsDTO);
 
-		execution.setSettings(settings);
-		
-		executionService.save(execution);
-		
-		flashMessageService.success(ra, "msg.settings.save.success");
+		flashMessageService.success(ra, Messages.EXECUTION_SAVE_SETTINGS_SUCCESS);
 
-		return "redirect:/execution/" + executionId;
+		return "redirect:/execution/" + executionId+"#settings";
 	}
 	
 	@PostMapping("/clear/user-feedback")
@@ -143,7 +142,7 @@ public class ExecutionController {
 			RedirectAttributes ra,
 			@PathVariable("executionId") String executionId) {
 			
-		Execution execution = executionService.findById(executionId);
+		Execution execution = executionService.findExecutionById(executionId);
 
 		for (NSolution<?> solution : execution.getSolutions()) {
 			SolutionUtils.clearUserFeedback(solution);
